Update DNS.storyboard
Removed "New DNS" title on the segue'd view and removed the large title option for a more clean look. Also linked the table view cell to also show the new DNS screen, with the intention of the information of the currently selected DNS listing to be inserted into the text fields to allow for editing and saving, to eliminate the need for multiple view controllers for both new DNS listings and editing existing ones.