Slight UI change, extra data addition
Added four more segments of DNS listing data to flesh out the tableView, as well as changed the dnsIdentifier to also include the "content" - aka IP Address of each listing, since when creating a new listing the IP address is a required data type so i figured it would make sense to show it as well. Had to modify constraints also.