Passed values to "Modify DNS Listings" Screen
Fixed the prepare to segue function that now correctly passes the information of the tapped table cell to the modify DNS listings screen, and also changed the identifier for the name into a label instead of a text field, as we only need to be able to edit the content/IP Address rather than the name of the DNS Listing also. Lastly, added 2 new functions into Cloudflare.swift, being updateDNS and newDNS. These will be called when the user is attempting to either save changes to their existing DNS Listing they are changing, or when saving a new DNS Listing for the first time. At first they will be simply throwing a true/false response depending on whether the API call (to be implemented in A2) is successful. If false is thrown then an alert will show, notifying the user.