Small fix, added pseudo to Cloudflare.swift
Added pesudocode to the 2 new functions in Cloudflare.swift file to dictate what will need to be done when it comes to implementing the Cloudflare API in assignment 2. I believe to implement the alert view in this class i will need to store the function for it in the "Modify DNS" and "New DNS" view controllers and call upon them via the Cloudflare.swift file as I can't seem to do it without importing more unnecessary stuff.