Final functionality
when the user clicks the save button in the nav bar for either the "edit dns" or "new dns" views; the app would check if the data has been successfully passed to the API (that we're implementing later?) so that could include incorrect formats of the ip address when changing, or names that exceed character limits, etc. If a true value is passed back into the function then the user simply goes back to the DNS tableView and can go about their day, but if there was an error and false is passed back; then they get the pictured error message (which we should be able to maybe customise depending on error? idk that part yet) then they can just click ok and try again

aside from the final piece of functionality, i mainly just cleaned up my code altogether here and made sure anything that needs to be implemented in assignment 2 is commented to some degree